Protect against trying to delete the md5 if it doesn't exist
When ftp fails to download a file after max number of tries, it will try to
delete the md5 file multiple times. This ensures it only deletes it once.

+++ pubmedparser/ftp.py
@@ -66,7 +66,8 @@ def _download_files(
                 print(f"{file_name} failed md5sum check max tries, deleting")
                 os.unlink(in_cache(file_name))
 
-        os.unlink(in_cache(md5_file_name))
+        if os.path.exists(md5_file_name):
+            os.unlink(in_cache(md5_file_name))
 
 
 def _list_local_pubmed_files(path: str, name_regex_template) -> List[str]:
